Added PostgreSQL job.
git-svn-id: https://svn.nixos.org/repos/nix/services/trunk@9832 70bd8c7a-acb8-0310-9f0d-9cc1c95dcdbb

+++ postgresql/builder.sh
@@ -8,6 +8,9 @@ echo "Creating cluster ctl script."
 cat >> $out/bin/control <<EOF
 #! $SHELL
 
+export PATH=$su/bin:$coreutils/bin
+export HOME=/homeless-shelter
+
 mkdir -p $logDir # !!! hack, should happen as $serveruser
 
 EXTRAOPTS=""
@@ -19,15 +22,17 @@ fi
 
 start()
 {
-    env -i LANG=C PGPORT="$port" $postgresql/bin/pg_ctl -D $dataDir \$EXTRAOPTS -w start
+    su - $serverUser -s /bin/sh -c 'env -i HOME=/homeless-shelter PATH=\$PATH LANG=C PGPORT="$port" $postgresql/bin/pg_ctl -D $dataDir \$EXTRAOPTS -w start'
 }
 
 stop()
 {
-    env -i LANG=C PGPORT="$port" $postgresql/bin/pg_ctl -D $dataDir \$EXTRAOPTS -w stop
+    su - $serverUser -s /bin/sh -c 'env -i HOME=/homeless-shelter PATH=\$PATH LANG=C PGPORT="$port" $postgresql/bin/pg_ctl -D $dataDir \$EXTRAOPTS -w stop'
 }
 
 if test "\$1" = start; then
+
+    trap stop 15
 
     start
 
@@ -36,19 +41,30 @@ elif test "\$1" = stop; then
     stop
 
 elif test "\$1" = init; then
+   
+    echo 'Are you sure you want to remove old PostgreSQL DBs?'
+    read
+    if ! test "\$REPLY" = yes; then exit 1; fi
 
-    env -i LANG=C $postgresql/bin/initdb -D $dataDir
+    rm -rf $dataDir
+    mkdir -p $dataDir
+    chown $serverUser $dataDir 
+    chmod 0700 $dataDir
+    su - $serverUser -s /bin/sh -c "env -i HOME=/homeless-shelter PATH=\$PATH LANG=C $postgresql/bin/initdb -D $dataDir"
+    touch $dataDir/postgresql.conf
 
     echo -n > $dataDir/pg_hba.conf
-    echo "local all all trust" >> $dataDir/pg_hba.conf
+    echo "local all all ${authMethod}" >> $dataDir/pg_hba.conf
     for i in $allowedHosts; do
         # !!! hack
         ip=\$(host \$i | sed 's/.* has address //')
-        echo "host all all \$ip/32 trust" >> $dataDir/pg_hba.conf
+        echo "host all all \$ip/32 ${authMethod}" >> $dataDir/pg_hba.conf
     done
-    
+   
     start
     
+    su - $serverUser -s /bin/sh -c "env -i HOME=/homeless-shelter PATH=\$PATH LANG=C $postgresql/bin/createuser -a -d root"
+
     for i in $subServices; do
         echo Subservice \$i...
         if test -f \$i/bin/control; then
